Fix naming in SFNT (#8)
* Fix naming in SFNT

[why]
The PS name and the UniqueID should match.

Copyright needs some value. I'm not sure if it should be removed.
Same goes with the Trademark field. But here we just erase it.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>

* Fixup: Use correct variable name

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>

* Remove spaces from PS name

[why]
PS font names may not contain blancs.
Due to the different naming (as compared to my setup/used on your blog)
the 'name' now 'args.name' contains blancs.

[how]
Remove the blancs from the PS name.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>