Keep all existing glyphs (no overwrite)
[why]
As pointed out by Jaykul there is the possibility that some glyphs that
are added by our script are already existing in Cascadia Code. These
probably fit better to the general font style than the generic glyphs we
patch into.

It would (maybe?) good not to overwrite pre-existing glyphs.
A more visible hint that something of this kind happens would be nice.

[how]
Use the 'careful' mode of font-patcher, that does not overwrite glyphs.
Insert a new processing step that just highlights skipped glyph patches.

Signed-off-by: Fini <ulf.fini.jastrow@desy.de>