Adapt to new Cascadia Code weight (#23)
* Use globbing to find patched file

[why]
When font-patcher patches a font the output font is always named after
the sourceFont.fullname, with "Nerd Font ..." added.

With the latest update Microsoft added the 'Regular' tag to the font (we
already patched that in always with the rename-font script btw). This
breaks our workflow because the expected hardcoded filename changed.

[how]
Unfortunately the font-patcher has no option to specify the output
filename.

Utilizing globbing the processing chain can be designed more robust.
This creates other risks of course, but in the intended use as git
workflow the directory will always be clean and thus work.

Signed-off-by: Fini <ulf.fini.jastrow@desy.de>

* Keep all existing glyphs (no overwrite)

[why]
As pointed out by Jaykul there is the possibility that some glyphs that
are added by our script are already existing in Cascadia Code. These
probably fit better to the general font style than the generic glyphs we
patch into.

It would (maybe?) good not to overwrite pre-existing glyphs.
A more visible hint that something of this kind happens would be nice.

[how]
Use the 'careful' mode of font-patcher, that does not overwrite glyphs.
Insert a new processing step that just highlights skipped glyph patches.

Signed-off-by: Fini <ulf.fini.jastrow@desy.de>