Fill sfntRevision with our version
[why]
We fill the PS version and the SFNT version strings with our
version/release number. But the SFNT revision is not updated.

This went unnoticed because we always want(ed) the same revision
number as already set by Microsoft.

[how]
The SFNT revision consists of two integers. Instead of setting them
directly somehow from our version string we let fontforge do the job of
guessing and fiddling.
If the sfntRevsison is unset fontforge will regenerate a new one based
on the other version strings.

[note]
Also output the version to the console so that we can check the git tag
mechanism easily. (The version is taken from the (last) git tag.)

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>