Allow lightweight tags to determine version
[why]
The font version is taken from the git tag of the release commit.
Releases should be tagged with an annotated tag, as only that is a
commit in itself, can be signed, ...

Release 1911.21 has been tagged with a lightweight tag only. That is
ignored and so the version we generated with that release was
v1910.04.1-2-g75828d2

[how]
To prevent wrong tagging take also lightweight tags into account.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>