Adapt to new Cascadia release scheme
[why]
Microsoft changed the release mechanism.
Instead of directly releasing the font files they are packed into a zip
archive now, in different formats.

Further complication is that the release version name is also in the zip
archive name, so we can not directly download it anymore in a generic
way.

[how]
First fetch the latest-release website and try to cut out the latest
release's directory and release package name.

Use that to download the archive. Unpack it and pull the font out.

This is all not so nice, but I see no other way to stay halfway generic
with their release scheme. And this just hopes, that they do not change
it again the next time :-/

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>