Use original PL glyphs if available
[why]
The powerline glyphs (blocks, triangulars) did not match up in hight
anymore after the last commit.

[how]
The reason is not known and not investigated. It could be a glitch in
the resizing code (which would have to be fixed, but that is another
repository).
A more elegant solution is to use the carefully crafted PL glyphs that
are already existing in the Microsoft supplied font.

So the CascadiaCodePL.ttf is used and we do not patch from scratch.
The --careful option ensure that we do not overwrite preexisting glyphs.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>