Fix Cascadia Code 2007.01 problem with font-patcher
[why]
With the Cascadia Code 2007.01 release the font-patcher process ends
with a segfault. The reason is unknown. Probably some internal structure
of the ttf file and the patching process leads to triggering a bug in
fontforge.

[how]
Nicola (xbb@github) found out, that rewriting the font file with fontforge
is possible. The rewriting process skips some of the font internal
properties - one of them is the reason for the crash.

So we open and save the Cascadia Code font without doing anything to it,
and work with that 'cleaned' font file onwards.

Thanks to Nicola for the hint with rewriting.

Fixes: #39

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>