Do not depend on exact font naming
[why]
font-patcher renames known fonts (and just adds some naming to unknown
fonts). Which source fonts it knows or not can change anytime.

We fixed the font-patcher version, but maybe want to update some time in
the future, so it would be good to not depend on the exact file name of
the output font.

[how]
- Put prepared fonts into additional directory.
- Use full name globbing, only depend on first letter (i.e. 'C')
- Remove intermediate font after renaming to Delugia*.ttf
- Repeat for all generated fonts

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>