Do not valign symbols in book font
[why]
When glyphs are added to a monospaced font and these glyphs are bigger
symbols (like on-off or logos) they need to be scaled down quite a bit
(for normal target fonts that are higher than wide).
These tiny symbols are then centered within the complete vertical hight
(descent to ascent) to have them centered on the powerline triangle
tips.

When a non-monospace font is generated the symbols are not tiny anymore
and often it is visually more pleasing to put them where they were
intended to be put relative to the baseline - because usually the source
fonts are well designed enough.

[how]
Do a vertical shift only on monospaced output fonts.

This is in accordance with the line 711 comment:
    # Now that we have copy/pasted the glyph, if we are creating a monospace
    # font we need to scale and move the glyphs.

-> Not scale and MOVE when non-monospaced.

Fixes: 53

Reported-by: Charles Roper <charles@roper.im>
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>