rename-font: Add option to specify style
[why]
The script is hardcoded to provide a 'Regular' font face. Now that we
want to generate other faces we need to be more flexible.

[how]
Add a parameter that is used in the appropriate places.

Sometimes the style has no additional name ("Italic" face). To allow
that we need to be careful with added blancs in the name strings.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>