Add light and bold (and italic) faces
[why]
The demand for correctly working bold and italic fonts can not be
satisfied with our simple approach.

We need a whole lot more fonts!

[how]
Use the static Cascadia font versions and do the patching on all
relevant fonts.

Use a helper script to invoke the actual patching to hide all the
parameters that stay the same for all the fonts.

Group the created fonts by interest (small (powerline), complete, book)
and release as zip packages.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>