Fix font versioning
[why]
The version in the font files reflects not the version name we want to
have. In fact it is just the git hash of the topmost commit and not the
git tag.

[how]
When switching to actions/checkout@v2 we only fetch the latest commit into
the git workflow (for performance reasons is this the default). But that
lacks all the tags, that we use to communicate the font version.

Now we resort to full depth cloning of the Delugia repository, which is
tiny anyway.

[note]
Action had been updated from @1 to @2 with commit
  22d4cae Combine workflows (#57)

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>