Change general font family naming scheme
[why]
We have 'normal' and 'bold' in one font family, but the 'italic' is in a
separate family. Some applications can group the weights/styles of one
font family in a way that allows to select 'bold' and/or 'italic' in the
application easily - without choosing a different font.

This seems to be broken if the italics is in a separate font family.

We have these families -> styles:

Deligia{}
  -> Regular, Bold
Delugia{}Light
  -> "" (Regular)
Delugia{}Italic
  -> Italic, Bold-Italic
Delugia{}ItalicLight
  -> Italic

[how]
Restructure the family names -> styles as such:

Deligia{}
  -> Regular, Bold, Italic, Bold-Italic
Delugia{}Light
  -> "" (Regular), Bold

[note]
With {} being our 5 styles:
  PL, MonoPL, Complete, MonoComplete, Book

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>