Create Light and Italic font faces (#59)
* Add gitignore file

[why]
It is easier to work with the scripts and git, if all normally created
intermediate files are ignored. It allows to do a 'git clean -f' for
example.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>

* rename-font: Correct help text

[why]
The version parameter is not optional, so it should not be noted as
optional.

I'm not sure if it ever was optional, maybe a leftover from a different,
previous approach.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>

* rename-font: Add option to specify style

[why]
The script is hardcoded to provide a 'Regular' font face. Now that we
want to generate other faces we need to be more flexible.

[how]
Add a parameter that is used in the appropriate places.

Sometimes the style has no additional name ("Italic" face). To allow
that we need to be careful with added blancs in the name strings.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>

* Add light and bold (and italic) faces

[why]
The demand for correctly working bold and italic fonts can not be
satisfied with our simple approach.

We need a whole lot more fonts!

[how]
Use the static Cascadia font versions and do the patching on all
relevant fonts.

Use a helper script to invoke the actual patching to hide all the
parameters that stay the same for all the fonts.

Group the created fonts by interest (small (powerline), complete, book)
and release as zip packages.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>

* remove prepare script

[why]
When we work with the static font files we do not need to prepare (fix
for working with them in fontforge) them anymore.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>

* Fix font versioning

[why]
The version in the font files reflects not the version name we want to
have. In fact it is just the git hash of the topmost commit and not the
git tag.

[how]
When switching to actions/checkout@v2 we only fetch the latest commit into
the git workflow (for performance reasons is this the default). But that
lacks all the tags, that we use to communicate the font version.

Now we resort to full depth cloning of the Delugia repository, which is
tiny anyway.

[note]
Action had been updated from @1 to @2 with commit
  22d4cae Combine workflows (#57)

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>

* Change general font family naming scheme

[why]
We have 'normal' and 'bold' in one font family, but the 'italic' is in a
separate family. Some applications can group the weights/styles of one
font family in a way that allows to select 'bold' and/or 'italic' in the
application easily - without choosing a different font.

This seems to be broken if the italics is in a separate font family.

We have these families -> styles:

Deligia{}
  -> Regular, Bold
Delugia{}Light
  -> "" (Regular)
Delugia{}Italic
  -> Italic, Bold-Italic
Delugia{}ItalicLight
  -> Italic

[how]
Restructure the family names -> styles as such:

Deligia{}
  -> Regular, Bold, Italic, Bold-Italic
Delugia{}Light
  -> "" (Regular), Bold

[note]
With {} being our 5 styles:
  PL, MonoPL, Complete, MonoComplete, Book

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>

* Fill the full name into fullname fields

[why]
The full name fields should contain the full name (obviously)
including the subfamily.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>

* Create more unique UniqueID

[why]
If one specific font is to be identified we need an ID that changes with
new revisions.

[how]
Add the version to the ID.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>

* Change to release in 5 archives

[why]
These Families are created:
- Delugia PL
- Delugia Mono PL
- Delugia Complete
- Delugia Mono Complete
- Delugia Book

At the moment we put both non-mono and mono families into one archive
like so:
1. Delugia PL and Delugia Mono PL
2. Delugia Complete and Delugia Mono Complete
3. Delugia Book

For some users it might make more sense to bundle the non-mono and mono
fonts. The majority probably needs just one family.

It might be unclear for users that these are 5 independent font sets.

[how]
Just stuff the files into more archives:
1. Delugia PL
2. Delugia Complete
3. Delugia Book

Ligarture-less variants:
4. Delugia Mono PL
5. Delugia Mono Complete

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>

* Whitespace cleanup

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>

* Correct license year

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>

* readme: Remove old news

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>

* readme: Add font sample images

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>