Do not add glyphs that meanwhile are present
[why]
Cascadia Code in the meanwhile has itself a more complete set of glyphs,
which include 0u2262, 0u2263, and 0u276F. So we do not need to rip them
out of Hack and put them into Delugia.
This also has the advantage (?) that the glyphs are slanted if a italic
font is used.

[how]
Just omit adding them.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>