Fix adding 0u26A1
[why]
With commit 5f152229
    Add 0u26A1 from Hack
we added (tried to add??!) the high voltage glyph to the basic
(non-complete) font.

Apparently that glyph is not in the font we use as source (Hack). Maybe
it fixed the issue at that time, because the user switched to our complete
version where it is added from another source.

[how]
Use a source font that has the glyph and patch that into the basic (non
complete) font files.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>