Improve skipped glyph diagnostics
[why]
We want to use the powerline symbols provided by Cascadia Code. The
normal font-patcher process wants to add these (and skips them
because we request non-overwriting with option -c).

So we expect some messages that glyphs are skipped, and we do not want
to see them in the DEBUG log (because they do not indicate possible
problems).

[how]
Filter the messages by (hopefully) the source font name and skip
Powerline*Symbol.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>