ci: Correct font version information for new Cascadia releases
[why]
When the CI is triggered (via cron) for a new Cascadia release we tag
our repository only after successful Delugia generation.

But we use the tag to determine which version is to be written into the
font files, that we create.

This means the version information in the font files is correct for
manually run release jobs, but not on an cron-automated one.

[how]
First impulse was to just move the tagging up to happen before we create
the new fonts - in fact before we fetch our repo?

But maybe we want to tag our repo only if we could successfully release
something.

To achieve this we use the repo's tag only if there was already a release
with the latest Cascadia version.
If the latest Cascadia version has never been release by us (as Delugia)
we directly use the new version number as our version.

If we succeed in creating the fonts our repo will get the tag anyhow.

Fixes: #66

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>