Fix font usability in Windows console apps
[why]
The generated fonts do not turn up in the font chooser of Windows CMD
and PowerShell (and probably more).

[how]
For some reasons Windows does not identify the fonts as being strictly
monospaced, so they are hidden in that font choosers.

For the monospaced fonts we set now the Panose proportion 'monospaced'.
Windows seems to honor the Panose properties.

It is not clear why we need to set the old Panose props, especially as
Cascadia Code does not (!) set them and is still detected as monospaced.

Anyhow, the way Windows detects if a font is monospaced is a mystery (at
least for me), and this works, so ;-)

Fixes: #68

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>