Correct naming behavior
[why]
The Light fonts are not grouped to the Regular and Bold fonts in
typographic names aware applications.

[how]
This is a longer standing problem with font-patcher, that does not fill
the typographic family and style with meaningful values. We could do
better with the `do_rename` script, but we did not.

Instead of repairing the `do_rename` script the `name_parser.py` module
is included, that I would like to get into `font-parser` to repair its
behaviour for all fonts.
As noone knows when or if at all it will end up in the patcher script,
we add the same code here.

The module will set the 4 names (family, subfamily, typogr. family and
typogr. style) in a way that we get correct grouping in both application
types: typographic aware one and 'ordinary' ones.

[note]
The `name_parser.py` originates here:
https://github.com/ryanoasis/nerd-fonts/pull/706

Fixes: #72

Reported-by: Rashil Gandhi <mail@rashil2000.me>
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>