CI: Select Ubuntu 22.04 as runner platform
[why]
Using a fixed platform prevents surprises when 'latest' is advanced to
the next version.

[how]
Use 22.04 as this is the latest available release as github runner.
Latest equals 20.04 at the moment.

22.04 is also choosen (very weak reason) because it has a bugfixed InkScape.
Admittedly we do not use it here. But Nerd Fonts has to use it for that
reason (for some jobs).

Need to also install fuse for the AppImage.

Suggested-by: Adam Cooper <mail@admcpr.com>
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>