Fix wrong font flags
[why]
The patched font looks different than the original Cascadia Cove font.
Character height seems to differ a bit, esp at smaller sizes.

[how]
With Fontforge it is not possible to control some aspects of the
generated font. There is no access to the ascpects from within
fontforge, and they are also not preserved but rather overwritten with
some default values. But these affect the font rendering engines.

Namely this are the flags and the lowest-recommended-ppem of the HEAD
table in the font.

The Nerd Fonts font-patcher addresses this shortcoming by manipulating
the HEAD table after the patching process directly.

But here we call (after running the font-patcher) fontforge again to
rename the font. In that process the details in the HEAD table are again
overwritten with stupid default values.

So we take the HEAD table manipulation code from font-patcher and call
it again after the renaming.

Differently from Nerd Fonts we copy the font flags exactly and do not
only clear the integer-ppem bit.

Fixes: #78

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>